Add files via upload
This is a repository of all projects i made. These are intermediate level of HTML, CSS and JS projects
task-2.1 : It is a form validation of a form using java script. It also includes creating and styling of the form using HTML and CSS.
task-2.2 : This project involves creating a responsive webpage using flex box , grid and media queries.
task-2.3 : This is a simple image gallery webpage which lets you to add images using url and remove images. Make sure to include the given font for more beautiful page title.